+++ packages/react-theme-provider/src/components/ThemeProvider/useThemeProvider.ts
@@ -39,8 +39,7 @@ export const useThemeProvider = (
   const theme = mergeThemes(parentTheme, state.theme ?? {});
   const themeClassName = useThemeStyleTag({ theme, targetDocument: state.targetDocument });
 
-  state.themeClassName = themeClassName;
   // ax is not needed here because `themeClassName` is not from a `makeStyles` call
-  state.className = [state.className || '', state.themeClassName].filter(Boolean).join(' ');
+  state.className = [state.className || '', themeClassName].filter(Boolean).join(' ');
   return state;
 };